+++ src/State.fs
@@ -1,6 +1,6 @@
 namespace State
 
-module ConcurentStorage =
+module ConcurrentStorage =
     open System.Collections.Concurrent
 
     type Key<'UniqueData> = Key of 'UniqueData